Fixes on top of the clang-warn-fix commit

+++ shlr/grub/include/grub/misc.h
@@ -42,7 +42,7 @@
 #endif
 
 #define ALIGN_UP(addr, align) \
-	((addr + (int) align - 1) & ~((int) align - 1))
+	((addr + (size_t) align - 1) & ~((size_t) align - 1))
 #define ARRAY_SIZE(array) (sizeof (array) / sizeof (array[0]))
 #define COMPILE_TIME_ASSERT(cond) switch (0) { case 1: case !(cond): ; }
